Add a customizable badge to the header
Prior to this change, the graylog in DEV mode provided
a badge in the header which helped developers to seperate
the GUI they are working on from production builds.
Users asked for something like this, so the can seperate
there graylog instances between production, staging and
testing.

This change adds a Cluster configuration which can be
changed in 'System -> Configuration'. You can enable
or disable the header badge, provide a 5 characters long
text and change the background color.