Allow remote debugging of graylog server
This allows to attach a remote debugger to the server container.
The biggest caveat is that the dynamic port to which the debugger can be
attached is only known after starting the container. Hence one would
have to set a breakpoint after the container start to pause execution,
identify the debug port and start the remote debugger with it.

- Added debug toggle via env var GRAYLOG_IT_DEBUG_SERVER
- Added optional debug build arg to container
- Added logging of debug port, if debugging is enabled
- Extracted building the container to NodeContainerFactory